rafs: use u32 to represent user io size
u32 should be enough.

+++ rafs/src/fs.rs
@@ -727,7 +727,7 @@ impl FileSystem for Rafs {
             let r = self.device.read_to(w, desc)?;
             result += r;
             recorder.mark_success(r);
-            if r != desc.bi_size {
+            if r as u32 != desc.bi_size {
                 break;
             }
         }
